# docs/openapi/public-openapi.json
{
  "openapi" : "3.1.0",
  "info" : {
    "title" : "x-ai-gateway Public API",
    "version" : "2026.05.21",
    "description" : "公开接入面的最小 OpenAPI 事实源，按头部 native provider 和 Lossless Translation Matrix 覆盖 docs、OpenAI-compatible、Claude/Gemini/Vertex 功能性入口、Codex Responses smoke 边界和 Media provider matrix；不可无损翻译的资源属性直接失败。"
  },
  "servers" : [ {
    "url" : "https://gateway.example.com",
    "description" : "生产网关地址示例"
  } ],
  "paths" : {
...
    "/v1/responses" : {
      "post" : {
        "summary" : "OpenAI-compatible Responses",
        "operationId" : "post_v1_responses",
        "tags" : [ "gateway" ],
        "security" : [ {
          "bearerAuth" : [ ]
        } ],
        "responses" : {
          "200" : {
            "description" : "OK"
          }
        },
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "type" : "object",
                "required" : [ "model" ],
                "properties" : {
                  "model" : {
                    "type" : "string",
                    "description" : "OpenAI model name or gateway model alias."
                  },
                  "input" : {
                    "type" : "object",
                    "description" : "Responses input string, message array or item array."
                  },
                  "stream" : {
                    "type" : "boolean",
                    "description" : "When true, emit OpenAI Responses-style server-sent events. OpenAI Direct streams pass through upstream raw SSE events when native passthrough is available."
                  },
                  "stream_options" : {
                    "type" : "object",
                    "description" : "Responses streaming options such as include_obfuscation; obfuscation is included on delta events by default and can be disabled with include_obfuscation=false."
                  },
                  "store" : {
                    "type" : "boolean",
                    "description" : "When true, persist the Response as a local resp_ resource."
                  },
                  "background" : {
                    "type" : "boolean",
                    "description" : "When true, mark stored local Response as background-capable for cancel."
                  },
                  "metadata" : {
                    "type" : "object",
                    "description" : "OpenAI metadata object preserved by the gateway."
                  },
                  "include" : {
                    "type" : "array",
                    "description" : "Additional output fields requested by OpenAI Responses clients."
                  },
                  "previous_response_id" : {
                    "type" : "string",
                    "description" : "Previous Response id for conversation continuation."
                  },
                  "tools" : {
                    "type" : "array",
                    "description" : "Responses tool definitions. Function tools execute through canonical tool calling; hosted file_search requires OpenAI Direct native execution and otherwise fails with native_hosted_tool_required; other hosted/MCP/custom tools are rejected explicitly instead of being ignored."
                  },
                  "tool_choice" : {
                    "type" : "object",
                    "description" : "Responses tool choice string or object. Non-function forced tool choices are rejected explicitly until their execution boundary is implemented."
                  }
                }
              }
            }
          }
        },
        "parameters" : [ {
          "name" : "Idempotency-Key",
          "in" : "header",
          "required" : false,
          "description" : "Optional key for non-streaming response replay.",
          "schema" : {
            "type" : "string"
          }
        } ]
      }
    },
...
    "/v1/responses/input_tokens" : {
      "post" : {
        "summary" : "Count Response input tokens",
        "operationId" : "post_v1_responses_input_tokens",
        "tags" : [ "gateway" ],
        "security" : [ {
          "bearerAuth" : [ ]
        } ],
        "responses" : {
          "200" : {
            "description" : "OK"
          }
        },
        "description" : "Counts Response input tokens. OpenAI Direct routes use native upstream counting when available; route-unavailable or non-native cases fail with native_input_tokens_required instead of returning a local deterministic estimate.",
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "type" : "object",
                "required" : [ "model" ],
                "properties" : {
                  "model" : {
                    "type" : "string",
                    "description" : "OpenAI model name or gateway model alias."
                  },
                  "input" : {
                    "type" : "object",
                    "description" : "Responses input string, message array or item array."
                  },
                  "stream" : {
                    "type" : "boolean",
                    "description" : "When true, emit OpenAI Responses-style server-sent events. OpenAI Direct streams pass through upstream raw SSE events when native passthrough is available."
                  },
                  "stream_options" : {
                    "type" : "object",
                    "description" : "Responses streaming options such as include_obfuscation; obfuscation is included on delta events by default and can be disabled with include_obfuscation=false."
                  },
                  "store" : {
                    "type" : "boolean",
                    "description" : "When true, persist the Response as a local resp_ resource."
                  },
                  "background" : {
                    "type" : "boolean",
                    "description" : "When true, mark stored local Response as background-capable for cancel."
                  },
                  "metadata" : {
                    "type" : "object",
                    "description" : "OpenAI metadata object preserved by the gateway."
                  },
                  "include" : {
                    "type" : "array",
                    "description" : "Additional output fields requested by OpenAI Responses clients."
                  },
                  "previous_response_id" : {
                    "type" : "string",
                    "description" : "Previous Response id for conversation continuation."
                  },
                  "tools" : {
                    "type" : "array",
                    "description" : "Responses tool definitions. Function tools execute through canonical tool calling; hosted file_search requires OpenAI Direct native execution and otherwise fails with native_hosted_tool_required; other hosted/MCP/custom tools are rejected explicitly instead of being ignored."
                  },
                  "tool_choice" : {
                    "type" : "object",
                    "description" : "Responses tool choice string or object. Non-function forced tool choices are rejected explicitly until their execution boundary is implemented."
                  }
                }
              }
            }
          }
        }
      }
    },
    "/v1/responses/compact" : {
      "post" : {
        "summary" : "Compact Response input context",
        "operationId" : "post_v1_responses_compact",
        "tags" : [ "gateway" ],
        "security" : [ {
          "bearerAuth" : [ ]
        } ],
        "responses" : {
          "200" : {
            "description" : "OK"
          }
        },
        "description" : "Compacts Response input context. OpenAI Direct routes use native upstream compaction when available; route-unavailable or non-native cases fail with native_compaction_required instead of returning a local opaque marker.",
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "type" : "object",
                "required" : [ "model" ],
                "properties" : {
                  "model" : {
                    "type" : "string",
                    "description" : "OpenAI model name or gateway model alias."
                  },
                  "input" : {
                    "type" : "object",
                    "description" : "Responses input string, message array or item array."
                  },
                  "stream" : {
                    "type" : "boolean",
                    "description" : "When true, emit OpenAI Responses-style server-sent events. OpenAI Direct streams pass through upstream raw SSE events when native passthrough is available."
                  },
                  "stream_options" : {
                    "type" : "object",
                    "description" : "Responses streaming options such as include_obfuscation; obfuscation is included on delta events by default and can be disabled with include_obfuscation=false."
                  },
                  "store" : {
                    "type" : "boolean",
                    "description" : "When true, persist the Response as a local resp_ resource."
                  },
                  "background" : {
                    "type" : "boolean",
                    "description" : "When true, mark stored local Response as background-capable for cancel."
                  },
                  "metadata" : {
                    "type" : "object",
                    "description" : "OpenAI metadata object preserved by the gateway."
                  },
                  "include" : {
                    "type" : "array",
                    "description" : "Additional output fields requested by OpenAI Responses clients."
                  },
                  "previous_response_id" : {
                    "type" : "string",
                    "description" : "Previous Response id for conversation continuation."
                  },
                  "tools" : {
                    "type" : "array",
                    "description" : "Responses tool definitions. Function tools execute through canonical tool calling; hosted file_search requires OpenAI Direct native execution and otherwise fails with native_hosted_tool_required; other hosted/MCP/custom tools are rejected explicitly instead of being ignored."
                  },
                  "tool_choice" : {
                    "type" : "object",
                    "description" : "Responses tool choice string or object. Non-function forced tool choices are rejected explicitly until their execution boundary is implemented."
                  }
                }
              }
            }
          }
        }
      }
    },
...
    "/v1/responses/{responseId}/cancel" : {
      "post" : {
        "summary" : "Cancel stored background Response",
        "operationId" : "post_v1_responses_responseId_cancel",
        "tags" : [ "gateway" ],
        "security" : [ {
          "bearerAuth" : [ ]
        } ],
        "responses" : {
          "200" : {
            "description" : "OK"
          }
        },
        "description" : "Cancels local background Responses or forwards cancel to the original OpenAI Direct upstream Response when lineage is available. Unknown remote resp_ ids require an explicit model route hint.",
        "parameters" : [ {
          "name" : "responseId",
          "in" : "path",
          "required" : true,
          "description" : "Stored Response id.",
          "schema" : {
            "type" : "string"
          }
        }, {
          "name" : "model",
          "in" : "query",
          "required" : false,
          "description" : "Gateway route hint for unknown remote OpenAI Direct resp_ ids that do not have local lineage.",
          "schema" : {
            "type" : "string"
          }
        }, {
          "name" : "X-AI-Gateway-OpenAI-Model",
          "in" : "header",
          "required" : false,
          "description" : "Gateway route hint header for unknown remote OpenAI Direct resp_ ids that do not have local lineage.",
          "schema" : {
            "type" : "string"
          }
        } ]
      }
    },
...
  },
  "components" : {
    "securitySchemes" : {
      "bearerAuth" : {
        "type" : "http",
        "scheme" : "bearer",
        "bearerFormat" : "Distributed Key"
      }
    }
  }
}